try to reduced duplicated code
